requirements.txt
requests
python-dotenv
psycopg2-binary
SQLAlchemy
pandas
fastapi
uvicorn